+++ hdate/htables.py
@@ -3,11 +3,10 @@
 import datetime as dt
 from dataclasses import dataclass
 from enum import Enum, IntEnum, auto
-from typing import Callable, TypeVar, Union
+from typing import Callable, Union
 
+from hdate.hebrew_date import CHANGING_MONTHS, LONG_MONTHS, Days, HebrewDate, Months
 from hdate.translator import TranslatorMixin
-
-HebrewDateT = TypeVar("HebrewDateT", bound="HebrewDate")  # type: ignore # noqa: F821
 
 
 def erange(start: Enum, end: Enum) -> list[Enum]:
@@ -352,57 +351,7 @@ PARASHA_SEQUENCES: dict[tuple[int, ...], tuple[Enum, ...]] = {
 }
 
 
-class Days(TranslatorMixin, IntEnum):
-    """Enum class for the days of the week."""
-
-    SUNDAY = 1
-    MONDAY = 2
-    TUESDAY = 3
-    WEDNESDAY = 4
-    THURSDAY = 5
-    FRIDAY = 6
-    SATURDAY = 7
-
-
-class Months(TranslatorMixin, IntEnum):
-    """Enum class for the Hebrew months."""
-
-    TISHREI = 1
-    MARCHESHVAN = 2
-    KISLEV = 3
-    TEVET = 4
-    SHVAT = 5
-    ADAR = 6
-    NISAN = 7
-    IYYAR = 8
-    SIVAN = 9
-    TAMMUZ = 10
-    AV = 11
-    ELUL = 12
-    ADAR_I = 13
-    ADAR_II = 14
-
-
-LONG_MONTHS = (
-    Months.TISHREI,
-    Months.SHVAT,
-    Months.ADAR_I,
-    Months.NISAN,
-    Months.SIVAN,
-    Months.AV,
-)
-SHORT_MONTHS = (
-    Months.TEVET,
-    Months.ADAR,
-    Months.ADAR_II,
-    Months.IYYAR,
-    Months.TAMMUZ,
-    Months.ELUL,
-)
-CHANGING_MONTHS = (Months.MARCHESHVAN, Months.KISLEV)
-
-
-def year_is_after(year: int) -> Callable[[HebrewDateT], bool]:
+def year_is_after(year: int) -> Callable[[HebrewDate], bool]:
     """
     Return a lambda function.
 
@@ -412,7 +361,7 @@ def year_is_after(year: int) -> Callable[[HebrewDateT], bool]:
     return lambda x: x.year > year
 
 
-def year_is_before(year: int) -> Callable[[HebrewDateT], bool]:
+def year_is_before(year: int) -> Callable[[HebrewDate], bool]:
     """
     Return a lambda function.
 
@@ -424,7 +373,7 @@ def year_is_before(year: int) -> Callable[[HebrewDateT], bool]:
 
 def move_if_not_on_dow(
     original: int, replacement: int, dow_not_orig: Days, dow_replacement: Days
-) -> Callable[[HebrewDateT], bool]:
+) -> Callable[[HebrewDate], bool]:
     """
     Return a lambda function.
 
@@ -437,7 +386,7 @@ def move_if_not_on_dow(
     )
 
 
-def correct_adar() -> Callable[[HebrewDateT], Union[bool, Callable[[], bool]]]:
+def correct_adar() -> Callable[[HebrewDate], Union[bool, Callable[[], bool]]]:
     """
     Return a lambda function.
 
@@ -451,7 +400,7 @@ def correct_adar() -> Callable[[HebrewDateT], Union[bool, Callable[[], bool]]]:
     )
 
 
-def not_rosh_chodesh() -> Callable[[HebrewDateT], bool]:
+def not_rosh_chodesh() -> Callable[[HebrewDate], bool]:
     """The 1st of Tishrei is not Rosh Chodesh."""
     return lambda x: not (x.month == Months.TISHREI and x.day == 1)
 
@@ -481,7 +430,7 @@ class Holiday(TranslatorMixin):
         tuple[Union[int, tuple[int, ...]], Union[Months, tuple[Months, ...]]], tuple[()]
     ]
     israel_diaspora: str
-    date_functions_list: list[Callable[[HebrewDateT], Union[bool, Callable[[], bool]]]]
+    date_functions_list: list[Callable[[HebrewDate], Union[bool, Callable[[], bool]]]]
 
 
 HOLIDAYS = (
